fix: admin 등록 위치를 feed앱으로 이동
